+++ server/migrations/sql/20250911002318_create_collection_table.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS collection
     account_number              VARCHAR(100) NOT NULL DEFAULT '',
     sender                      VARCHAR(255) NOT NULL DEFAULT 'Anonymous',
     date                        VARCHAR(100) NOT NULL DEFAULT '',
-    amount                      INT NOT NULL DEFAULT 0,
+    amount                      FLOAT        NOT NULL DEFAULT 0,
     currency                    VARCHAR(100) NOT NULL DEFAULT '',
     created_at                  TIMESTAMP    DEFAULT current_timestamp,
     created_by                  VARCHAR(100) NOT NULL DEFAULT '',
